Update example to use PascalVOC loader
