+++ examples/models/object_detection/retina_net/basic/pascal_voc/train.py
@@ -50,74 +50,32 @@ if FLAGS.wandb_entity:
 """
 ## Data loading
 
-First, let's define the data-loading function: `load_pascal_voc()`.  KerasCV supports a
-`bounding_box_format` argument in all components that process bounding boxes.  To match
-the KerasCV API style, it is recommended that when writing a data loader, you support a
-`bounding_box_format` argument.  This makes it clear to those invoking your
-data loader what format the bounding boxes are in.
+In this guide, we use the data-loading function: `keras_cv.loaders.pascal_voc.load()`.
+KerasCV supports a `bounding_box_format` argument in all components that process
+bounding boxes.  To match the KerasCV API style, it is recommended that when writing a
+custom data loader, you also support a `bounding_box_format` argument.
+This makes it clear to those invoking your data loader what format the bounding boxes
+are in.
 
 For example:
 
 ```python
-train_ds = load_pascal_voc(split='train', bounding_box_format='xywh', batch_size=8)
+train_ds, ds_info = keras_cv.loaders.pascal_voc.load(split='train', bounding_box_format='xywh', batch_size=8)
 ```
 
 Clearly yields bounding boxes in the format `xywh`.  You can read more about
 KerasCV bounding box formats [in the API docs](https://keras.io/api/keras_cv/bounding_box/formats/).
-"""
 
-
-def curry_map_function(bounding_box_format, img_size):
-    """Mapping function to create batched image and bbox coordinates"""
-
-    resizing = keras.layers.Resizing(
-        height=img_size[0], width=img_size[1], crop_to_aspect_ratio=False
-    )
-
-    def apply(inputs):
-        inputs["image"] = resizing(inputs["image"])
-        inputs["objects"]["bbox"] = bounding_box.convert_format(
-            inputs["objects"]["bbox"],
-            images=inputs["image"],
-            source="rel_yxyx",
-            target=bounding_box_format,
-        )
-
-        bounding_boxes = inputs["objects"]["bbox"]
-        labels = tf.cast(inputs["objects"]["label"], tf.float32)
-        labels = tf.expand_dims(labels, axis=-1)
-        bounding_boxes = tf.concat([bounding_boxes, labels], axis=-1)
-        return {"images": inputs["image"], "bounding_boxes": bounding_boxes}
-
-    return apply
-
-
-def load_pascal_voc(
-    split, bounding_box_format, batch_size, shuffle=True, img_size=(512, 512)
-):
-    dataset, dataset_info = tfds.load(
-        "voc/2007", split=split, shuffle_files=shuffle, with_info=True
-    )
-    dataset = dataset.map(
-        curry_map_function(bounding_box_format=bounding_box_format, img_size=img_size),
-        num_parallel_calls=tf.data.AUTOTUNE,
-    )
-    dataset = dataset.shuffle(8 * batch_size)
-    dataset = dataset.apply(
-        tf.data.experimental.dense_to_ragged_batch(batch_size=batch_size)
-    )
-    return dataset, dataset_info
-
-
-"""
-Great!  Our data is now loaded into the format
+Our data comesloaded into the format
 `{"images": images, "bounding_boxes": bounding_boxes}`.  This format is supported in all
 KerasCV preprocessing components.
 
 Lets load some data and verify that our data looks as we expect it to.
 """
 
-dataset, _ = load_pascal_voc(split="train", bounding_box_format="xywh", batch_size=9)
+dataset, _ = keras_cv.loaders.pascal_voc.load(
+    split="train", bounding_box_format="xywh", batch_size=9
+)
 
 
 def visualize_dataset(dataset, bounding_box_format):
@@ -160,10 +118,10 @@ friendly data augmentation inside of a `tf.data` pipeline.
 
 # train_ds is batched as a (images, bounding_boxes) tuple
 # bounding_boxes are ragged
-train_ds, _ = load_pascal_voc(
+train_ds, _ = keras_cv.loaders.pascal_voc.load(
     bounding_box_format="xywh", split="train", batch_size=FLAGS.batch_size
 )
-val_ds, _ = load_pascal_voc(
+val_ds, _ = keras_cv.loaders.pascal_voc.load(
     bounding_box_format="xywh", split="validation", batch_size=FLAGS.batch_size
 )
